test(EuiDataGrid): move EuiDataGrid updating column definitions jest tests to a separate file
