+++ packages/eui/src/components/datagrid/data_grid_test_utils.tsx
@@ -1,0 +1,179 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0 and the Server Side Public License, v 1; you may not use this file except
+ * in compliance with, at your election, the Elastic License 2.0 or the Server
+ * Side Public License, v 1.
+ */
+
+import { ReactWrapper } from 'enzyme';
+import type { EuiDataGridProps, RenderCellValue } from './data_grid_types';
+import { findTestSubject } from '../../test';
+import { act } from '@testing-library/react';
+
+export function extractGridData(datagrid: ReactWrapper<EuiDataGridProps>) {
+  const rows: string[][] = [];
+
+  const headerCells = findTestSubject(datagrid, 'dataGridHeaderCell', '|=');
+  const headerRow: string[] = [];
+  headerCells.forEach((cell: any) =>
+    headerRow.push(cell.find('div.euiDataGridHeaderCell__content').text())
+  );
+  rows.push(headerRow);
+
+  // reduce the virtualized grid of cells into rows
+  const columnCount = datagrid.prop('columnVisibility').visibleColumns.length;
+  const gridCells = findTestSubject(datagrid, 'dataGridRowCell');
+  const visibleRowsCount = gridCells.length / columnCount;
+  for (let i = 0; i < visibleRowsCount; i++) {
+    const rowContent: string[] = [];
+    for (let j = i * columnCount; j < (i + 1) * columnCount; j++) {
+      const cell = gridCells.at(j);
+      rowContent.push(cell.find('[data-test-subj="cell-content"]').text());
+    }
+    rows.push(rowContent);
+  }
+
+  return rows;
+}
+
+export function openColumnSorterSelection(datagrid: ReactWrapper) {
+  let columnSelectionPopover = datagrid.find(
+    'EuiPopover[data-test-subj="dataGridColumnSortingPopoverColumnSelection"]'
+  );
+  expect(columnSelectionPopover).not.euiPopoverToBeOpen();
+  act(() => {
+    columnSelectionPopover
+      .find('button[data-test-subj="dataGridColumnSortingSelectionButton"]')
+      .simulate('click');
+  });
+
+  datagrid.update();
+
+  columnSelectionPopover = datagrid.find(
+    'EuiPopover[data-test-subj="dataGridColumnSortingPopoverColumnSelection"]'
+  );
+  expect(columnSelectionPopover).euiPopoverToBeOpen();
+
+  return columnSelectionPopover;
+}
+
+export function closeColumnSorterSelection(datagrid: ReactWrapper) {
+  let columnSelectionPopover = datagrid.find(
+    'EuiPopover[data-test-subj="dataGridColumnSortingPopoverColumnSelection"]'
+  );
+  // popover will go away if all of the columns are selected
+  if (columnSelectionPopover.length > 0) {
+    expect(columnSelectionPopover).euiPopoverToBeOpen();
+
+    act(() => {
+      columnSelectionPopover
+        .find('button[data-test-subj="dataGridColumnSortingSelectionButton"]')
+        .simulate('click');
+    });
+
+    datagrid.update();
+
+    columnSelectionPopover = datagrid.find(
+      'EuiPopover[data-test-subj="dataGridColumnSortingPopoverColumnSelection"]'
+    );
+    expect(columnSelectionPopover).not.euiPopoverToBeOpen();
+  }
+
+  return columnSelectionPopover;
+}
+
+export function openColumnSorter(datagrid: ReactWrapper) {
+  let popover = datagrid.find(
+    'EuiPopover[data-test-subj="dataGridColumnSortingPopover"]'
+  );
+  expect(popover).not.euiPopoverToBeOpen();
+
+  act(() => {
+    popover
+      .find('button[data-test-subj="dataGridColumnSortingButton"]')
+      .simulate('click');
+  });
+
+  datagrid.update();
+
+  popover = datagrid.find(
+    'EuiPopover[data-test-subj="dataGridColumnSortingPopover"]'
+  );
+  expect(popover).euiPopoverToBeOpen();
+
+  return popover;
+}
+
+export function closeColumnSorter(datagrid: ReactWrapper) {
+  let popover = datagrid.find(
+    'EuiPopover[data-test-subj="dataGridColumnSortingPopover"]'
+  );
+  expect(popover).euiPopoverToBeOpen();
+
+  act(() => {
+    popover
+      .find('button[data-test-subj="dataGridColumnSortingButton"]')
+      .simulate('click');
+  });
+
+  datagrid.update();
+
+  popover = datagrid.find(
+    'EuiPopover[data-test-subj="dataGridColumnSortingPopover"]'
+  );
+  expect(popover).not.euiPopoverToBeOpen();
+
+  return popover;
+}
+
+export function openColumnSelector(datagrid: ReactWrapper) {
+  let popover = datagrid.find(
+    'EuiPopover[data-test-subj="dataGridColumnSelectorPopover"]'
+  );
+  expect(popover).not.euiPopoverToBeOpen();
+
+  act(() => {
+    popover
+      .find('button[data-test-subj="dataGridColumnSelectorButton"]')
+      .simulate('click');
+  });
+
+  datagrid.update();
+
+  popover = datagrid.find(
+    'EuiPopover[data-test-subj="dataGridColumnSelectorPopover"]'
+  );
+  expect(popover).euiPopoverToBeOpen();
+
+  return popover;
+}
+
+export function closeColumnSelector(datagrid: ReactWrapper) {
+  let popover = datagrid.find(
+    'EuiPopover[data-test-subj="dataGridColumnSelectorPopover"]'
+  );
+  expect(popover).euiPopoverToBeOpen();
+
+  act(() => {
+    popover
+      .find('button[data-test-subj="dataGridColumnSelectorButton"]')
+      .simulate('click');
+  });
+
+  datagrid.update();
+
+  popover = datagrid.find(
+    'EuiPopover[data-test-subj="dataGridColumnSelectorPopover"]'
+  );
+  expect(popover).not.euiPopoverToBeOpen();
+
+  return popover;
+}
+
+export const renderCellValueRowAndColumnCount: RenderCellValue = ({
+  rowIndex,
+  columnId,
+}) => `${rowIndex}, ${columnId}`;
+
+export const renderCellRowAsValue: RenderCellValue = ({ rowIndex }) => rowIndex;
